RAG_Eval_Training_Krishna/requirements.txt
python-dotenv>=1.0.1
pypdf>=4.2.0
rich>=13.7.1
pandas>=2.2.2

# Vector store
chromadb>=0.5.5
langchain-chroma>=0.1.4

# LangChain core + providers
langchain>=0.2.16
langchain-community>=0.2.16
langchain-openai>=0.1.22
langchain-anthropic>=0.1.23

# RAG evaluation
ragas>=0.1.12
datasets>=2.20.0
